chore: tidy base runner imports
