+++ espnet3/inference/__init__.py
@@ -1,3 +1,4 @@
-from .abs_metrics import *
-from .inference_runner import *
-from .score_runner import *
+from .abs_metrics import *  # noqa: F401,F403
+from espnet3.runner import BaseRunner, InferenceRunner, ScoreRunner
+
+__all__ = [name for name in globals() if not name.startswith("_")]
